--- shortlog
@@ -1,15 +1,0 @@
-Device:  cuda:0
-    Epoch: [1/10] Step: [10/258]  loss: 0.15  acc: 0.78  tp: 0.12  tn: 0.67  fp: 0.13  fn: 0.08 
-    Epoch: [1/10] Step: [20/258]  loss: 0.10  acc: 0.87  tp: 0.15  tn: 0.73  fp: 0.07  fn: 0.05 
-    Epoch: [1/10] Step: [30/258]  loss: 0.09  acc: 0.88  tp: 0.13  tn: 0.76  fp: 0.06  fn: 0.05 
-    Epoch: [1/10] Step: [40/258]  loss: 0.08  acc: 0.89  tp: 0.11  tn: 0.79  fp: 0.03  fn: 0.07 
-    Epoch: [1/10] Step: [50/258]  loss: 0.07  acc: 0.91  tp: 0.16  tn: 0.76  fp: 0.05  fn: 0.04 
-    Epoch: [1/10] Step: [60/258]  loss: 0.08  acc: 0.88  tp: 0.10  tn: 0.79  fp: 0.07  fn: 0.04 
-    Epoch: [1/10] Step: [70/258]  loss: 0.07  acc: 0.90  tp: 0.13  tn: 0.77  fp: 0.03  fn: 0.06 
-    Epoch: [1/10] Step: [80/258]  loss: 0.07  acc: 0.89  tp: 0.17  tn: 0.73  fp: 0.06  fn: 0.04 
-    Epoch: [1/10] Step: [90/258]  loss: 0.06  acc: 0.93  tp: 0.08  tn: 0.85  fp: 0.04  fn: 0.03 
-    Epoch: [1/10] Step: [100/258]  loss: 0.06  acc: 0.92  tp: 0.16  tn: 0.76  fp: 0.03  fn: 0.04 
-    Epoch: [1/10] Step: [110/258]  loss: 0.06  acc: 0.92  tp: 0.12  tn: 0.80  fp: 0.04  fn: 0.04 
-    Epoch: [1/10] Step: [120/258]  loss: 0.06  acc: 0.91  tp: 0.17  tn: 0.75  fp: 0.03  fn: 0.05 
-    Epoch: [1/10] Step: [130/258]  loss: 0.06  acc: 0.91  tp: 0.14  tn: 0.78  fp: 0.02  fn: 0.07 
-    Epoch: [1/10] Step: [140/258]  loss: 0.06  acc: 0.91  tp: 0.12  tn: 0.80  fp: 0.05  fn: 0.03
